Revalidation of hypothesis for calibrated Consciousness Level for
the Great Books of the Western World in Power vs. Force being a
average value of the individual calibrated Consciousness Levels of the Books of the Western World. Hypothesis validated as untrue.
Also validated hypothesis that the calibrated Consciousness Level of the Great Books of the Western World equals the median value of the individual Great Books of the Western World. Hypothesis validated as untrue.

Also verified that the used functions work.